Sistema de conversão de vídeos com codec invalido

+++ pro/med.php
@@ -12,32 +12,41 @@ if ($med){ # Se a média existir
         
         if ($ac=='eliminar'){ # Se a ação for eliminar
         
-            $caminho = "/home/guilha/www/media.drena.xyz/";
+            if ($med['est']!='2'){ # Se o estado da média não for 2 (processando)
+                $caminho = "/home/guilha/www/media.drena.xyz/";
+    
+                $caminho_ori = $caminho.'ori/'.$med['id'].'.'.$med_ext; # Original
+                unlink($caminho_ori);
+            
+                $caminho_som = $caminho.'som/'.$med['id'].'.'.$med_ext; # Som
+                unlink($caminho_som);
+            
+                $caminho_img = $caminho.'img/'.$med['id'].'.'.$med_ext; # Imagem
+                unlink($caminho_img);
+    
+                $caminho_webm = $caminho.'webm/'.$med['id'].'.webm';    # Comprimido
+                unlink($caminho_webm);
 
-            $caminho_ori = $caminho.'ori/'.$med['id'].'.'.$med_ext; # Original
-            unlink($caminho_ori);
-        
-            $caminho_som = $caminho.'som/'.$med['id'].'.'.$med_ext; # Som
-            unlink($caminho_som);
-        
-            $caminho_img = $caminho.'img/'.$med['id'].'.'.$med_ext; # Imagem
-            unlink($caminho_img);
-
-            $caminho_webm = $caminho.'webm/'.$med['id'].'.webm';    # Processado
-            unlink($caminho_webm);
-        
-            $caminho_thumb = $caminho.'thumb/'.$med['thu'].'.jpg';  # Thumb
-            unlink($caminho_thumb);
-
-            # Se existir algum dos ficheiros que supostamente foram apagados
-            if (file_exists($caminho_ori) OR file_exists($caminho_som) OR file_exists($caminho_img) OR file_exists($caminho_webm) OR file_exists($caminho_thumb)){
-                echo "Erro: Não foi possivel remover os ficheiros.";
-            } else if ($bd->query("DELETE FROM med_gos WHERE med='".$med['id']."'") === FALSE) {
-                echo "Erro: ".$bd->error;
-            } else if ($bd->query("DELETE FROM med WHERE id='".$med['id']."'") === FALSE) {
-                echo "Erro: ".$bd->error;
+                $caminho_convertido = $caminho.'conv/'.$med['id'].'.webm'; # Convertido
+                unlink($caminho_convertido);
+            
+                $caminho_thumb = $caminho.'thumb/'.$med['thu'].'.jpg';  # Thumb
+                unlink($caminho_thumb);
+    
+                # Se existir algum dos ficheiros que supostamente foram apagados
+                if (file_exists($caminho_ori) OR file_exists($caminho_som) OR file_exists($caminho_img) OR file_exists($caminho_webm) OR file_exists($caminho_convertido) OR file_exists($caminho_thumb)){
+                    echo "Erro: Não foi possivel remover os ficheiros.";
+                } else if ($bd->query("DELETE FROM med_com WHERE med='".$med['id']."'") === FALSE) {
+                    echo "Erro mysql: ".$bd->error;
+                } else if ($bd->query("DELETE FROM med_gos WHERE med='".$med['id']."'") === FALSE) {
+                    echo "Erro mysql: ".$bd->error;
+                } else if ($bd->query("DELETE FROM med WHERE id='".$med['id']."'") === FALSE) {
+                    echo "Erro mysql: ".$bd->error;
+                } else {
+                    header("Location: /../perfil?uti=".$uti['nut']);
+                }
             } else {
-                header("Location: /../perfil?uti=".$uti['nut']);
+                echo "Erro: Não podes eliminar a média enquanto está a ser processada.";
             }
             exit;
         
@@ -54,7 +63,7 @@ if ($med){ # Se a média existir
         } else if ($ac=='comprimir'){ # Alterar título da média
             
             if ($med['tip']=='1' AND $med['est']=='1'){ # Se a media for um vídeo e o estado for 1 (bitrate alto)
-                $output = exec("php /home/guilha/www/drena.xyz/pro/med_compressao.php ".$med['id']." > /dev/null &");
+                exec("php /home/guilha/www/drena.xyz/pro/med_compressao.php ".$med['id']." > /dev/null &");
                 sleep(2);
                 header("Location: ".$_SERVER['HTTP_REFERER']);
             } else {
